backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth>=2.2.5
django-allauth==0.59.0
django-cors-headers==4.4.0
django-multiselectfield==0.1.13

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-decouple==3.8
PyJWT==2.9.0
redis==5.2.0
sqlparse==0.5.1
serpapi
google-search-results
